Move, but do not refactor, printers to their own package
